eat_archive/requirements.txt
Django==1.3
django-tinymce==1.5.1b2
geopy==0.94.2
networkx==1.6
image==1.5.27